requirements.txt
# Document parsing
pdfplumber>=0.10.0
python-docx>=1.1.0
python-pptx>=0.6.23

# Audio transcription
openai-whisper>=20231117

# Text-to-Speech (for audio output) - OPTIONAL
# Note: TTS requires Python <3.12. If using Python 3.12+, TTS features will be disabled
# and the system will fall back to SRT subtitle generation for audio files.
# To install TTS (Python 3.9-3.11 only): pip install TTS soundfile scipy
# soundfile>=0.12.1  # Uncomment if using TTS
# scipy>=1.11.0      # Uncomment if using TTS

# Translation
httpx>=0.27.0

# PDF reconstruction
reportlab>=4.2.0

# API server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# UI
streamlit>=1.35.0
requests>=2.32.0